fix some frontend pages and add email validation

+++ src/main/java/com/example/resourciumoptima2/service/EmployeeService.java
@@ -2,12 +2,12 @@ package com.example.resourciumoptima2.service;
 
 import com.example.resourciumoptima2.entity.Employee;
 import com.example.resourciumoptima2.repository.EmployeeRepo;
-import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.Persistence;
 
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class EmployeeService {
     private final EmployeeRepo employeeRepo;
@@ -18,11 +18,19 @@ public class EmployeeService {
         this.employeeRepo = new EmployeeRepo(entityManagerFactory);
     }
 //    add Employee
-    public void addEmlpoyee(Employee employee) {
+    public boolean addEmlpoyee(Employee employee) {
+        Employee employee1 = (Employee) employeeRepo.auth(employee);
+        if(employee1!=null || !validateEmail(employee.getEmail())){
+            return true;
+        }
         employeeRepo.saveEmployee(employee);
+        return false;
     }
 //    login
     public Object userLogin(Employee employee){
+        if(employee.getUserName().isEmpty()){
+            return null;
+        }
       return employeeRepo.auth(employee);
     }
 //    get All Employees
@@ -31,7 +39,19 @@ public class EmployeeService {
         return employeeList;
     }
 
-    public void deleteEmp(String userId) {
+    public boolean deleteEmp(String userId) {
+        if(userId.isEmpty()){
+            return false;
+        }
         employeeRepo.deleteEmployee(userId);
+        return true;
+    }
+
+    public void updateProfile(Employee employee) {
+        employeeRepo.update_Profile(employee);
+    }
+    public boolean validateEmail(String email){
+            String regex = "^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$";
+            return email.matches(regex);
     }
 }
